requirements.txt
azure-storage-blob
matplotlib
numpy
pandas
pycaret
pycaret[models]
sktime
streamlit